moves all previous files in v1.x
