+++ v1.x/drop_tables.sql
