+++ v1.x/exporters/abstractexporter.py
