+++ v1.x/parsers/discogsartistparser.py
