+++ v1.x/parsers/discogslabelparser.py
