+++ v1.x/parsers/discogsmasterparser.py
