+++ v1.x/tools/make_artist_name_id_pair.py
